Split filter repository out
