+++ FootballHistory/Server/Controllers/LeagueSeasonController.cs
@@ -8,46 +8,49 @@ namespace FootballHistory.Server.Controllers
     [Route("api/[controller]")]
     public class LeagueSeasonController : Controller
     {
-        private ILeagueSeasonRepository Repository { get; }
-        public LeagueSeasonController(ILeagueSeasonRepository repository)
+        private ILeagueSeasonRepository LeagueSeasonRepository { get; }
+        private IFilterRepository FilterRepository { get; }
+
+        public LeagueSeasonController(ILeagueSeasonRepository leagueSeasonRepository, IFilterRepository filterRepository)
         {
-            Repository = repository;
+            LeagueSeasonRepository = leagueSeasonRepository;
+            FilterRepository = filterRepository;
         }
 
         [HttpGet("[action]")]
         public FilterOptions GetFilterOptions()
         {
-            return Repository.GetFilterOptions();
+            return FilterRepository.GetFilterOptions();
         }
         
         [HttpGet("[action]")]
         public DefaultFilter GetDefaultFilter()
         {
-            return Repository.GetDefaultFilter();
+            return FilterRepository.GetDefaultFilter();
         }
                 
         [HttpGet("[action]")]
         public List<ResultMatrixRow> GetResultMatrix(string tier, string season)
         {
-            return Repository.GetResultMatrix(Convert.ToInt32(tier), season);
+            return LeagueSeasonRepository.GetResultMatrix(Convert.ToInt32(tier), season);
         }
                                 
         [HttpGet("[action]")]
         public PlayOffs GetPlayOffMatches(string tier, string season)
         {
-            return Repository.GetPlayOffMatches(Convert.ToInt32(tier), season);
+            return LeagueSeasonRepository.GetPlayOffMatches(Convert.ToInt32(tier), season);
         }
                 
         [HttpGet("[action]")]
         public List<LeagueTableRow> GetLeagueTable(string tier, string season)
         {
-            return Repository.GetLeagueTable(Convert.ToInt32(tier), season);
+            return LeagueSeasonRepository.GetLeagueTable(Convert.ToInt32(tier), season);
         }
                 
         [HttpGet("[action]")]
         public LeagueRowDrillDown GetDrillDown(string tier, string season, string team)
         {
-            return Repository.GetDrillDown(Convert.ToInt32(tier), season, team);
+            return LeagueSeasonRepository.GetDrillDown(Convert.ToInt32(tier), season, team);
         }
     }
 }
