+++ FootballHistory/Server/Repositories/IFilterRepository.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using static FootballHistory.Server.Repositories.LeagueSeasonRepository;
+
+namespace FootballHistory.Server.Repositories
+{
+    public interface IFilterRepository
+    {
+        DefaultFilter GetDefaultFilter();
+        FilterOptions GetFilterOptions();
+    }
+}
