+++ FootballHistory/Server/Repositories/ILeagueSeasonRepository.cs
@@ -5,8 +5,6 @@ namespace FootballHistory.Server.Repositories
 {
     public interface ILeagueSeasonRepository
     {
-        DefaultFilter GetDefaultFilter();
-        FilterOptions GetFilterOptions();
         List<ResultMatrixRow> GetResultMatrix(int tier, string season);
         PlayOffs GetPlayOffMatches(int tier, string season);
         List<LeagueTableRow> GetLeagueTable(int tier, string season);
